GICv3: use edge flow handler for LPIs
When hierarchy interrupt domain is enabled, the flow handler of
the top domain which is closest to CPU will probably be used as
the handler of that linux interrupt.

However the MBIGEN in P660 implementation needs an acknowledgement
so that the level kept for MBIGEN interrupt can be cleared.

To achieve this, we simply let ITS driver use handle_edge_irq()
to deal with LPIs instead of handle_fasteoi_irq(), and issue an
explicit EOI for LPIs in ISR.

Signed-off-by: Yun Wu <wuyun.wu@huawei.com>
Signed-off-by: Ding Tianhong <dingtianhong@huawei.com>